+++ scripts/prepdocs/files/listfilestrategy.py
@@ -1,0 +1,153 @@
+import base64
+import hashlib
+import os
+import re
+import tempfile
+from abc import ABC
+from glob import glob
+from typing import IO, AsyncGenerator, Dict, List, Optional, Union
+
+from azure.core.credentials_async import AsyncTokenCredential
+from azure.storage.filedatalake.aio import (
+    DataLakeServiceClient,
+)
+
+
+class File:
+    def __init__(self, content: IO, acls: Optional[dict[str, list]] = None):
+        self.content = content
+        self.acls = acls or {}
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self):
+        self.content.close()
+
+    def filename(self):
+        return os.path.basename(self.content.name)
+
+    def filename_to_id(self):
+        filename_ascii = re.sub("[^0-9a-zA-Z_-]", "_", self.content.name)
+        filename_hash = base64.b16encode(self.content.name.encode("utf-8")).decode("ascii")
+        return f"file-{filename_ascii}-{filename_hash}"
+
+
+class ListFileStrategy(ABC):
+    async def list(self) -> AsyncGenerator[File, None]:
+        if False:
+            yield
+
+    async def list_paths(self) -> AsyncGenerator[str, None]:
+        if False:
+            yield
+
+
+class LocalListFileStrategy(ListFileStrategy):
+    def __init__(self, path_pattern: str, verbose: bool = False):
+        self.path_pattern = path_pattern
+        self.verbose = verbose
+
+    async def list_paths(self) -> AsyncGenerator[str, None]:
+        async for p in self._list_paths(self.path_pattern):
+            yield p
+
+    async def _list_paths(self, path_pattern: str) -> AsyncGenerator[str, None]:
+        for path in glob(path_pattern):
+            if os.path.isdir(path):
+                async for p in self._list_paths(f"{path}/*"):
+                    yield p
+
+            yield path
+
+    async def list(self) -> AsyncGenerator[File, None]:
+        async for path in self.list_paths():
+            if not self.check_md5(path):
+                yield File(content=open(path, mode="rb"))
+
+    def check_md5(self, path: str) -> bool:
+        # if filename ends in .md5 skip
+        if path.endswith(".md5"):
+            return True
+
+        # if there is a file called .md5 in this directory, see if its updated
+        stored_hash = None
+        with open(path, "rb") as file:
+            existing_hash = hashlib.md5(file.read()).hexdigest()
+        hash_path = f"{path}.md5"
+        if os.path.exists(hash_path):
+            with open(hash_path, encoding="utf-8") as md5_f:
+                stored_hash = md5_f.read()
+
+        if stored_hash and stored_hash.strip() == existing_hash.strip():
+            if self.verbose:
+                print(f"Skipping {path}, no changes detected.")
+                return True
+
+        # Write the hash
+        with open(hash_path, "w", encoding="utf-8") as md5_f:
+            md5_f.write(existing_hash)
+
+        return False
+
+
+class ADLSGen2ListFileStrategy(ListFileStrategy):
+    def __init__(
+        self,
+        data_lake_storage_account: str,
+        data_lake_filesystem: str,
+        data_lake_path: str,
+        credential: Union[AsyncTokenCredential, str],
+    ):
+        self.data_lake_storage_account = data_lake_storage_account
+        self.data_lake_filesystem = data_lake_filesystem
+        self.data_lake_path = data_lake_path
+        self.credential = credential
+
+    async def list_paths(self) -> AsyncGenerator[str, None]:
+        async with DataLakeServiceClient(
+            account_url=f"https://{self.data_lake_storage_account}.dfs.core.windows.net", credential=self.credential
+        ) as service_client, service_client.get_file_system_client(self.data_lake_filesystem) as filesystem_client:
+            async for path in filesystem_client.get_paths(path=self.data_lake_path, recursive=True):
+                if path.is_directory:
+                    continue
+
+                yield path.name
+
+    async def list(self) -> AsyncGenerator[File, None]:
+        async with DataLakeServiceClient(
+            account_url=f"https://{self.data_lake_storage_account}.dfs.core.windows.net", credential=self.credential
+        ) as service_client, service_client.get_file_system_client(self.data_lake_filesystem) as filesystem_client:
+            async for path in self.list_paths():
+                temp_file_path = os.path.join(tempfile.gettempdir(), os.path.basename(path))
+                try:
+                    async with filesystem_client.get_file_client(path) as file_client:
+                        with open(temp_file_path, "wb") as temp_file:
+                            downloader = await file_client.download_file()
+                            await downloader.readinto(temp_file)
+                    # Parse out user ids and group ids
+                    acls: Dict[str, List[str]] = {"oids": [], "groups": []}
+                    # https://learn.microsoft.com/python/api/azure-storage-file-datalake/azure.storage.filedatalake.datalakefileclient?view=azure-python#azure-storage-filedatalake-datalakefileclient-get-access-control
+                    # Request ACLs as GUIDs
+                    access_control = await file_client.get_access_control(upn=False)
+                    acl_list = access_control["acl"]
+                    # https://learn.microsoft.com/azure/storage/blobs/data-lake-storage-access-control
+                    # ACL Format: user::rwx,group::r-x,other::r--,user:xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx:r--
+                    acl_list = acl_list.split(",")
+                    for acl in acl_list:
+                        acl_parts: list = acl.split(":")
+                        if len(acl_parts) != 3:
+                            continue
+                        if len(acl_parts[1]) == 0:
+                            continue
+                        if acl_parts[0] == "user" and "r" in acl_parts[2]:
+                            acls["oids"].append(acl_parts[1])
+                        if acl_parts[0] == "group" and "r" in acl_parts[2]:
+                            acls["groups"].append(acl_parts[1])
+                    yield File(content=open(temp_file_path, "rb"), acls=acls)
+                except Exception as e:
+                    print(f"\tGot an error while reading {path} -> {e} --> skipping file")
+                    try:
+                        os.remove(temp_file_path)
+                    except Exception as e:
+                        print(f"\tGot an error while deleting {temp_file_path} -> {e}")
