+++ scripts/prepdocs/files/pdfparser.py
@@ -1,0 +1,113 @@
+import html
+from abc import ABC
+from typing import IO, AsyncGenerator, Union
+
+from azure.ai.formrecognizer import DocumentTable
+from azure.ai.formrecognizer.aio import DocumentAnalysisClient
+from azure.core.credentials import AzureKeyCredential
+from azure.core.credentials_async import AsyncTokenCredential
+from pypdf import PdfReader
+
+from prepdocs.strategy import USER_AGENT
+
+
+class Page:
+    def __init__(self, page_num: int, offset: int, text: str):
+        self.page_num = page_num
+        self.offset = offset
+        self.text = text
+
+
+class PdfParser(ABC):
+    async def parse(self, content: IO) -> AsyncGenerator[Page, None]:
+        if False:
+            yield
+
+
+class LocalPdfParser(PdfParser):
+    async def parse(self, content: IO) -> AsyncGenerator[Page, None]:
+        reader = PdfReader(content)
+        pages = reader.pages
+        offset = 0
+        for page_num, p in enumerate(pages):
+            page_text = p.extract_text()
+            yield Page(page_num=page_num, offset=offset, text=page_text)
+            offset += len(page_text)
+
+
+class DocumentAnalysisPdfParser(PdfParser):
+    def __init__(
+        self,
+        endpoint: str,
+        credential: Union[AsyncTokenCredential, AzureKeyCredential],
+        model_id="prebuilt-layout",
+        verbose: bool = False,
+    ):
+        self.model_id = model_id
+        self.endpoint = endpoint
+        self.credential = credential
+        self.verbose = verbose
+
+    async def parse(self, content: IO) -> AsyncGenerator[Page, None]:
+        if self.verbose:
+            print(f"Extracting text from '{content.name}' using Azure Form Recognizer")
+
+        async with DocumentAnalysisClient(
+            endpoint=self.endpoint, credential=self.credential, headers={"x-ms-useragent": USER_AGENT}
+        ) as form_recognizer_client:
+            poller = await form_recognizer_client.begin_analyze_document(model_id=self.model_id, document=content)
+            form_recognizer_results = await poller.result()
+
+            offset = 0
+            for page_num, page in enumerate(form_recognizer_results.pages):
+                tables_on_page = [
+                    table
+                    for table in (form_recognizer_results.tables or [])
+                    if table.bounding_regions and table.bounding_regions[0].page_number == page_num + 1
+                ]
+
+                # mark all positions of the table spans in the page
+                page_offset = page.spans[0].offset
+                page_length = page.spans[0].length
+                table_chars = [-1] * page_length
+                for table_id, table in enumerate(tables_on_page):
+                    for span in table.spans:
+                        # replace all table spans with "table_id" in table_chars array
+                        for i in range(span.length):
+                            idx = span.offset - page_offset + i
+                            if idx >= 0 and idx < page_length:
+                                table_chars[idx] = table_id
+
+                # build page text by replacing characters in table spans with table html
+                page_text = ""
+                added_tables = set()
+                for idx, table_id in enumerate(table_chars):
+                    if table_id == -1:
+                        page_text += form_recognizer_results.content[page_offset + idx]
+                    elif table_id not in added_tables:
+                        page_text += DocumentAnalysisPdfParser.table_to_html(tables_on_page[table_id])
+                        added_tables.add(table_id)
+
+                yield Page(page_num=page_num, offset=offset, text=page_text)
+                offset += len(page_text)
+
+    @classmethod
+    def table_to_html(cls, table: DocumentTable):
+        table_html = "<table>"
+        rows = [
+            sorted([cell for cell in table.cells if cell.row_index == i], key=lambda cell: cell.column_index)
+            for i in range(table.row_count)
+        ]
+        for row_cells in rows:
+            table_html += "<tr>"
+            for cell in row_cells:
+                tag = "th" if (cell.kind == "columnHeader" or cell.kind == "rowHeader") else "td"
+                cell_spans = ""
+                if cell.column_span is not None and cell.column_span > 1:
+                    cell_spans += f" colSpan={cell.column_span}"
+                if cell.row_span is not None and cell.row_span > 1:
+                    cell_spans += f" rowSpan={cell.row_span}"
+                table_html += f"<{tag}{cell_spans}>{html.escape(cell.content)}</{tag}>"
+            table_html += "</tr>"
+        table_html += "</table>"
+        return table_html
