+++ tests/test_prepdocs.py
@@ -4,12 +4,12 @@ import openai
 import pytest
 import tenacity
 from conftest import MockAzureCredential
-from scripts.prepdocs.files import (
-    ADLSGen2ListFileStrategy,
+from scripts.prepdocs.files.blobmanager import File
+from scripts.prepdocs.files.embeddings import (
     AzureOpenAIEmbeddingService,
-    File,
     OpenAIEmbeddingService,
 )
+from scripts.prepdocs.files.listfilestrategy import ADLSGen2ListFileStrategy
 
 
 def test_filename_to_id():
